used extention method with extension method we can seperate the logic just like we seperate the end point in the node with route
